rustling watch passed

+++ exercises/variables/variables6.rs
@@ -3,8 +3,9 @@
 // Execute `rustlings hint variables6` or use the `hint` watch subcommand for a
 // hint.
 
+// I AM DONE
 
-const NUMBER: i32 = 3;
+const NUMBER:u16 = 3;
 fn main() {
     println!("Number {}", NUMBER);
 }
